+++ src/main/java/tornadofx/Async.kt
@@ -11,6 +11,7 @@ import javafx.concurrent.Worker
 import javafx.scene.Node
 import javafx.scene.control.Labeled
 import javafx.scene.control.ProgressIndicator
+import javafx.scene.control.ToolBar
 import javafx.scene.layout.BorderPane
 import javafx.scene.layout.Region
 import javafx.util.Duration
@@ -236,10 +237,13 @@ fun <T : Any> Node.runAsyncWithProgress(progress: Node = ProgressIndicator(), op
             }
         }
     } else {
+        println("Parent of $this is $parent")
         val paddingHorizontal = (this as? Region)?.paddingHorizontal?.toDouble() ?: 0.0
         val paddingVertical = (this as? Region)?.paddingVertical?.toDouble() ?: 0.0
         (progress as? Region)?.setPrefSize(boundsInParent.width - paddingHorizontal, boundsInParent.height - paddingVertical)
-        val children = requireNotNull(parent.getChildList()) { "This node has no child list, and cannot contain the progress node" }
+        // Unwrap ToolBar parent, it has an extra HBox or VBox inside it, we need to target the items list
+        val p = (parent?.parent as? ToolBar) ?: parent
+        val children = requireNotNull(p.getChildList()) { "This node has no child list, and cannot contain the progress node" }
         val index = children.indexOf(this)
         children.add(index, progress)
         removeFromParent()