+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 
+- runAsyncWithProgress must Unwrap ToolBar parent (https://github.com/edvin/tornadofx/issues/687)
 - Calling `close()` inside `InternalWindow` would also close the parent
 - `placeholder` for list type controls should not require list type parameter
 - Added warning log message when `WorkspaceApp` is called with a `Workspace` subclass as main view paramter